Create SolidityTest_21BCS5650
In this I have created a token named "Happy" with token abbreviation as "hpy".The contract maps addresses to balances. The mint function takes two parameters as address and a value. The contract has a burn function which will function opposite to the mint function. It will deduct the value from the totalsupply.